requirements.txt
streamlit
openai
chromadb
pandas
python-dotenv
beautifulsoup4
requests
lxml
